+++ lib/bob_web/endpoint.ex
@@ -35,7 +35,6 @@ defmodule BobWeb.Endpoint do
 
   plug(Plug.RequestId)
   plug(Logster.Plugs.Logger, excludes: [:params])
-  plug(BobWeb.Plugs.Secret, api_only: true)
 
   plug(Plug.Parsers,
     parsers: [:urlencoded, :json, Bob.Plug.Parser],
